Strip the vendor string from XRE_main call to prevent using a vendor subdir for the profile.
Introducing this now would be a profile migration nightmare. Even if the back-end/build system supports this now, we need to ignore any vendor string that is set in application.ini or the static include of the nsAppData struct for mail client use.
This keeps our profile location at %APPDATA%/FossaMail (%APPDATA%/{APP_NAME})